@@ -13,16 +13,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.4",
-    "@angular/cdk": "^13.2.4",
-    "@angular/common": "^13.2.4",
-    "@angular/compiler": "^13.2.4",
-    "@angular/core": "^13.2.4",
-    "@angular/forms": "^13.2.4",
-    "@angular/material": "^13.2.4",
-    "@angular/platform-browser": "^13.2.4",
-    "@angular/platform-browser-dynamic": "^13.2.4",
-    "@angular/router": "^13.2.4",
+    "@angular/animations": "^14.1.3",
+    "@angular/cdk": "^14.1.3",
+    "@angular/common": "^14.1.3",
+    "@angular/compiler": "^14.1.3",
+    "@angular/core": "^14.1.3",
+    "@angular/forms": "^14.1.3",
+    "@angular/material": "^14.1.3",
+    "@angular/platform-browser": "^14.1.3",
+    "@angular/platform-browser-dynamic": "^14.1.3",
+    "@angular/router": "^14.1.3",
     "@ngx-translate/core": "^10.0.2",
     "base-x": "3.0.5",
     "bignumber.js": "^7.2.1",
@@ -35,21 +35,21 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.5",
-    "@angular-eslint/builder": "13.1.0",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/schematics": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "^13.2.5",
-    "@angular/compiler-cli": "^13.2.4",
-    "@angular/language-service": "^13.2.4",
+    "@angular-devkit/build-angular": "^14.1.3",
+    "@angular-eslint/builder": "14.0.2",
+    "@angular-eslint/eslint-plugin": "14.0.2",
+    "@angular-eslint/eslint-plugin-template": "14.0.2",
+    "@angular-eslint/schematics": "14.0.2",
+    "@angular-eslint/template-parser": "14.0.2",
+    "@angular/cli": "^14.1.3",
+    "@angular/compiler-cli": "^14.1.3",
+    "@angular/language-service": "^14.1.3",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.11.0",
-    "@typescript-eslint/parser": "5.11.0",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "eslint": "^8.18.0",
     "eslint-plugin-import": "2.25.4",
     "eslint-plugin-jsdoc": "37.9.4",
     "eslint-plugin-prefer-arrow": "1.2.3",
@@ -64,6 +64,6 @@
     "karma-read-json": "^1.1.0",
     "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "typescript": "~4.5.5"
+    "typescript": "~4.7.4"
   }
 }
