@@ -1,6 +1,6 @@
 import { Component, Inject } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { FormGroup, FormBuilder, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder, Validators } from '@angular/forms';
 
 import { HwWalletService } from '../../../../services/hw-wallet.service';
 import { ChildHwDialogParams } from '../hw-options-dialog/hw-options-dialog.component';
@@ -17,7 +17,7 @@ import { HwDialogBaseComponent } from '../hw-dialog-base.component';
   styleUrls: ['./hw-restore-seed-dialog.component.scss'],
 })
 export class HwRestoreSeedDialogComponent extends HwDialogBaseComponent<HwRestoreSeedDialogComponent> {
-  form: FormGroup;
+  form: UntypedFormGroup;
   // If true, the seed entered by the user will not be loaded on the device, the operation will
   // just compare it to the one on the device.
   justCheckingSeed: boolean;
@@ -26,7 +26,7 @@ export class HwRestoreSeedDialogComponent extends HwDialogBaseComponent<HwRestor
     @Inject(MAT_DIALOG_DATA) public data: ChildHwDialogParams,
     public dialogRef: MatDialogRef<HwRestoreSeedDialogComponent>,
     private hwWalletService: HwWalletService,
-    formBuilder: FormBuilder,
+    formBuilder: UntypedFormBuilder,
   ) {
     super(hwWalletService, dialogRef);
 
