@@ -1,7 +1,7 @@
 import { throwError as observableThrowError, SubscriptionLike, of } from 'rxjs';
 import { retryWhen, delay, mergeMap, debounceTime } from 'rxjs/operators';
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { BigNumber } from 'bignumber.js';
 import { HttpErrorResponse } from '@angular/common/http';
 
@@ -111,7 +111,7 @@ export class FormSourceSelectionComponent implements OnInit, OnDestroy {
   }
 
   sourceSelectionModes = SourceSelectionModes;
-  form: FormGroup;
+  form: UntypedFormGroup;
   // All available wallets.
   allWallets: WalletWithBalance[];
   // Wallet selected by the user.
@@ -140,7 +140,7 @@ export class FormSourceSelectionComponent implements OnInit, OnDestroy {
 
   constructor(
     private appService: AppService,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private balanceAndOutputsService: BalanceAndOutputsService,
   ) { }
 
