@@ -1,6 +1,6 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA, MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import { FormBuilder } from '@angular/forms';
+import { UntypedFormBuilder } from '@angular/forms';
 
 import { OfflineDialogsBaseComponent, OfflineDialogsStates } from '../offline-dialogs-base.component';
 import { copyTextToClipboard } from '../../../../../utils/general-utils';
@@ -51,7 +51,7 @@ export class CopyRawTxComponent extends OfflineDialogsBaseComponent implements O
     @Inject(MAT_DIALOG_DATA) private data: CopyRawTxData,
     public dialogRef: MatDialogRef<CopyRawTxComponent>,
     private msgBarService: MsgBarService,
-    formBuilder: FormBuilder,
+    formBuilder: UntypedFormBuilder,
   ) {
     super(formBuilder);
 
