@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 
 import { ButtonComponent } from '../../../layout/button/button.component';
 
@@ -37,7 +37,7 @@ export class OfflineDialogsBaseComponent {
   @ViewChild('okButton') okButton: ButtonComponent;
   // Allows to deactivate the form while the component is busy.
   working = false;
-  form: FormGroup;
+  form: UntypedFormGroup;
   currentState: OfflineDialogsStates = OfflineDialogsStates.Loading;
   states = OfflineDialogsStates;
   validateForm = false;
@@ -61,7 +61,7 @@ export class OfflineDialogsBaseComponent {
   dropdownElements: OfflineDialogsDropdownElement[];
 
   constructor(
-    _formBuilder: FormBuilder,
+    _formBuilder: UntypedFormBuilder,
   ) {
     this.form = _formBuilder.group({
       dropdown: [''],
